requirements.txt
tqdm
python-dotenv
langsmith
fastapi
python-dotenv
pydantic
pandas
groq==0.4.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-groq==0.0.1
numpy==1.26.4
pydantic==2.6.2
python-dotenv==1.0.1
tqdm==4.66.2
fastapi==0.104.0
uvicorn==0.24.0
uvicorn[standard]==0.24.0
python-multipart